requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
Django==5.0.2
django-crispy-forms==2.1
django-filter==23.5
djangorestframework==3.14.0
h11==0.14.0
idna==3.6
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pycparser==2.21
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
